+++ tests/TestCaseResultTest.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace MichaelHall\Webunit\Tests;
 
-use DataTypes\FilePath;
-use DataTypes\Url;
+use DataTypes\Net\Url;
+use DataTypes\System\FilePath;
 use MichaelHall\Webunit\Assertions\AssertContains;
 use MichaelHall\Webunit\AssertResult;
 use MichaelHall\Webunit\Location\FileLocation;