+++ tests/TestSuiteTest.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace MichaelHall\Webunit\Tests;
 
-use DataTypes\FilePath;
-use DataTypes\Url;
+use DataTypes\Net\Url;
+use DataTypes\System\FilePath;
 use MichaelHall\HttpClient\HttpClient;
 use MichaelHall\HttpClient\HttpClientInterface;
 use MichaelHall\Webunit\Assertions\AssertContains;